fix(temporal): checkpoint failure watch retries
